% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{final_death_comp_df}
\alias{final_death_comp_df}
\title{Create a final death per age dataset}
\usage{
final_death_comp_df(res, df)
}
\arguments{
\item{res}{squire model run}

\item{df}{Excess deaths per age per province}
}
\description{
Create a final death per age dataset
}
