% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{final_death_comp_province}
\alias{final_death_comp_province}
\title{Create a plot comparing final deaths per province per age from model to observed}
\usage{
final_death_comp_province(res, df)
}
\arguments{
\item{res}{squire model run}

\item{df}{Excess deaths per age per province}
}
\description{
Create a plot comparing final deaths per province per age from model to observed
}
