% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{final_deaths}
\alias{final_deaths}
\title{Create a final death per age dataset}
\usage{
final_deaths(res)
}
\arguments{
\item{res}{squire model run}
}
\description{
Create a final death per age dataset
}
