% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_utils.R
\name{format_sero_df}
\alias{format_sero_df}
\title{Format output of seroprev output from}
\usage{
format_sero_df(out)
}
\arguments{
\item{out}{Seroprev output from [[seroprev_df_det]]}
}
\description{
Format output of seroprev output from
}
