% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_utils.R
\name{generate_draws}
\alias{generate_draws}
\title{Generate draws using parameters drawn from posterior}
\usage{
generate_draws(
  out,
  draws = 10,
  parallel = TRUE,
  burnin = 100,
  log_likelihood = india_log_likelihood
)
}
\arguments{
\item{out}{Output of [[squire::pmcmc]]}

\item{draws}{How many draws are being used from pars_list. Default = NULL,
which will use all the pars.}

\item{parallel}{Are we simulating in parallel. Default = FALSE}

\item{pars_list}{Output of [[generate_parameters]]}

\item{interventions}{Are new interventions being used or default. Default = NULL}
}
\description{
Generate draws using parameters drawn from posterior
}
