% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_utils.R
\name{generate_parameters}
\alias{generate_parameters}
\title{Generate parameter draws from a squire pmcmc run}
\usage{
generate_parameters(out, draws = 10, burnin = 1000, ll = TRUE)
}
\arguments{
\item{out}{Output of [[squire::pmcmc]]}

\item{draws}{Number of draws from mcmc chain. Default = 10}
}
\description{
Generate parameter draws from a squire pmcmc run
}
