% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimue_utils.R
\name{nim_sq_format}
\alias{nim_sq_format}
\title{Nimue and squire format wrapper for conistent return}
\usage{
nim_sq_format(
  out,
  var_select = NULL,
  reduce_age = TRUE,
  combine_compartments = TRUE,
  date_0 = NULL
)
}
\arguments{
\item{out}{nimue_simulation object}

\item{var_select}{Vector of compartment names, e.g. \code{c("S", "R")}. In
  addition a number of summary compartment can be requested. These include:
\itemize{
      \item{"deaths"}{ Daily Deaths }
      \item{"infections"}{ Daily Infections }
      \item{"hospital_occupancy"}{ Occupied Hospital Beds }
      \item{"ICU_occupancy"}{ Occupied ICU Beds }
      \item{"hospital_demand"}{ Required Hospital Beds }
      \item{"ICU_demand"}{ Required ICU Beds }
      \item{"hospital_incidence"}{ Incidence of hospitilisation }
      \item{"ICU_incidence"}{ Incidence of ICU admissions }
      \item{"hospitilisations"}{ Incidence of hospital+ICU admissions }
      \item{""vaccines", "unvaccinated", "vaccinated", "priorvaccinated"}{ Vaccine outputs }
      }}

\item{reduce_age}{Collapse age-dimension, calculating the total in the
compartment.}

\item{combine_compartments}{Collapse compartments of same type together
(e.g. E1 and E2 -> E)}

\item{date_0}{Date of time 0, if specified a date column will be added}
}
\description{
Nimue and squire format wrapper for conistent return
}
