% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_utils.R
\name{plot_sero_df}
\alias{plot_sero_df}
\title{Plot seroprev curve}
\usage{
plot_sero_df(out, color = NULL, ci = TRUE)
}
\arguments{
\item{out}{Formatted seroprev output from [[format_sero_df]]}

\item{color}{Name of column that determines color of curves. Default = NULL}

\item{ci}{Logical for whether CI is plotted. Default = FALSE}
}
\description{
Plot seroprev curve
}
