% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_utils.R
\name{sero_det}
\alias{sero_det}
\title{Create a sero detection probability density function}
\usage{
sero_det(
  ret = "igg",
  igg_conv = 13.3,
  igm_conv = 12.3,
  igg_sens = 0.914,
  igm_sens = 0.894,
  igg_scale = 139,
  igm_scale = 50,
  pdfl = 600
)
}
\arguments{
\item{ret}{What antibody are we detecting. Must be one of "igg", "igm" or
"iggm". Default = "igg}

\item{igg_conv}{Mean days for IgG seroconversion. Default = 13.3}

\item{igm_conv}{Mean days for IgM seroconversion. Default = 12.3}

\item{igg_sens}{Max sensitivity for IgG. Default = 0.914}

\item{igm_sens}{Max sensitivity for IgM. Default = 0.894}

\item{igg_scale}{Weibull seroreversion decay scale for IgG. Default = 139}

\item{igm_scale}{Weibull seroreversion decay scale for IgM. Default = 50}

\item{pdfl}{Length of probability density function. Default = 600 days}
}
\description{
Create a sero detection probability density function
}
