% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sero_utils.R
\name{seroprev_df_det}
\alias{seroprev_df_det}
\title{Create a seropositivity data frame from squire run output}
\usage{
seroprev_df_det(res, sero_det)
}
\arguments{
\item{res}{squire model run}

\item{sero_det}{sero detection curve from [[sero_det]]}
}
\description{
Create a seropositivity data frame from squire run output
}
